test(gcp): fold pending/ tests into per-service sdk-* modules
Split the 4 active pending tests (Datastore, Logging, GCS gRPC v2, Managed Kafka)
into one-service-per-module sdk-* modules and remove the legacy pending/ module.
Wires the 4 modules into CI. Managed Kafka's test stays gcp_pending-gated here
since that service (PR #34) is not yet merged.
