+++ .github/workflows/ci.yml
@@ -158,7 +158,7 @@ jobs:
       - name: Download modules
         run: |
           go mod download
-          for m in sdk sdkv1 sdk-firestore sdk-monitoring sdk-workflows sdk-dataproc pending; do
+          for m in sdk sdkv1 sdk-firestore sdk-monitoring sdk-workflows sdk-dataproc sdk-datastore sdk-logging sdk-gcs-grpc sdk-managed-kafka; do
             (cd tests/integration/gcp/$m && go mod download)
           done
 
@@ -219,12 +219,31 @@ jobs:
           MONITORING_EMULATOR_HOST: localhost:8081
         run: go test -v -count=1 -timeout 120s ./...
 
-      - name: Run Logging/Datastore/GCS-gRPC SDK tests
-        working-directory: tests/integration/gcp/pending
+      - name: Run Datastore SDK tests
+        working-directory: tests/integration/gcp/sdk-datastore
+        env:
+          DATASTORE_EMULATOR_HOST: localhost:8081
+          GCP_EMULATOR_PROJECT: test-project
+        run: go test -v -count=1 -timeout 120s ./...
+
+      - name: Run Logging SDK tests
+        working-directory: tests/integration/gcp/sdk-logging
+        env:
+          LOGGING_EMULATOR_HOST: localhost:8081
+          GCP_EMULATOR_PROJECT: test-project
+        run: go test -v -count=1 -timeout 120s ./...
+
+      - name: Run GCS gRPC SDK tests
+        working-directory: tests/integration/gcp/sdk-gcs-grpc
         env:
           STORAGE_EMULATOR_HOST_GRPC: localhost:8081
-          DATASTORE_EMULATOR_HOST: localhost:8081
-          LOGGING_EMULATOR_HOST: localhost:8081
+          GCP_EMULATOR_PROJECT: test-project
+        run: go test -v -count=1 -timeout 120s ./...
+
+      - name: Run Managed Kafka SDK tests
+        working-directory: tests/integration/gcp/sdk-managed-kafka
+        env:
+          GCP_EMULATOR_ENDPOINT: http://localhost:8080/
           GCP_EMULATOR_PROJECT: test-project
         run: go test -v -count=1 -timeout 120s ./...
