+++ tests/integration/gcp/sdk-managed-kafka/go.mod
@@ -1,0 +1,11 @@
+module jaiscloud-gcp-sdk-managed-kafka-test
+
+go 1.26
+
+require github.com/stretchr/testify v1.11.1
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
