+++ tests/integration/gcp/sdk-managed-kafka/kafka_test.go
@@ -1,4 +1,4 @@
-package tests
+package kafka_test
 
 import (
 	"bytes"
